implement KinDynWrapper library
