+++ src/CMakeLists.txt
@@ -10,3 +10,4 @@ add_subdirectory(libSimplified_model_controllers)
 add_subdirectory(libRobot_helper)
 add_subdirectory(libWhole_body_controllers)
 add_subdirectory(libTrajectory_planner)
+add_subdirectory(libKinDyn_wrapper)
